Validando cadastros e visualizacoes pelo login de acesso. Mudança na visualização de onde ocorre erro.

+++ src/projeto/dao/DaoUsuarioRepository.java
@@ -4,12 +4,12 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
 import projeto.connection.SingleConnectionBD;
 import projeto.model.ModelLogin;
+import projeto.util.Utilitario;
 
 public class DaoUsuarioRepository {
 	private Connection connection;
@@ -19,77 +19,191 @@ public class DaoUsuarioRepository {
 		connection = SingleConnectionBD.getConnection();
 	}
 
-	public boolean gravarUsuario(ModelLogin ml) throws Exception {
+	public boolean gravarUsuario(ModelLogin ml, Long userIdLogado) throws Exception {
 		boolean gravado = false;
-		String sql = "insert into " + TABELA + " (nome, email, login, senha) values (?, ?, ?, ?)";
+		String sql = "insert into " + TABELA + " (nome, email, login, senha, usuario_id) values (?, ?, ?, ?, ?)";
 		PreparedStatement ps = null;
-		try {
-			ps = connection.prepareStatement(sql);
-			ps.setString(1, ml.getNome());
-			ps.setString(2, ml.getEmail());
-			ps.setString(3, ml.getLogin());
-			ps.setString(4, ml.getSenha());
-			ps.execute();
-			connection.commit();
-			gravado = true;
-		} catch (SQLException e) {
-			e.printStackTrace();
-			connection.rollback();
-		} finally {
-			if (ps != null)
-				ps.close();
+		if (ml != null && userIdLogado != null) {
+			try {
+				ps = connection.prepareStatement(sql);
+				ps.setString(1, ml.getNome());
+				ps.setString(2, ml.getEmail());
+				ps.setString(3, ml.getLogin());
+				ps.setString(4, ml.getSenha());
+				ps.setLong(5, userIdLogado);
+				ps.execute();
+				connection.commit();
+				gravado = true;
+			} catch (SQLException e) {
+				Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
+				e.printStackTrace();
+				connection.rollback();
+			} finally {
+				if (ps != null)
+					ps.close();
+			}
 		}
 
 		return gravado;
 	}
 
-	public ModelLogin obterUsuario(String login) throws Exception {
+	public ModelLogin obterUsuario(String login, Long idUserLogado) throws Exception {
 		ModelLogin ml = null;
-		String sql = "select * from " + TABELA + " where login = ?"; // login é único
+		String sql = "select * from " + TABELA + " where login = ? and usuario_id = ? and useradmin is false"; // login
+																												// é
+																												// único
 		PreparedStatement ps = null;
 		ResultSet rs = null;
-		try {
-			ps = connection.prepareStatement(sql);
-			ps.setString(1, login);
-			rs = ps.executeQuery();
-			while (rs.next()) {
-				ml = new ModelLogin();
-				ml.setId(rs.getLong("id"));
-				ml.setNome(rs.getString("nome"));
-				ml.setEmail(rs.getString("email"));
-				ml.setLogin(rs.getString("login"));
-				ml.setSenha(rs.getString("senha"));
+		if (login != null && !login.isEmpty() && idUserLogado != null) {
+			try {
+				ps = connection.prepareStatement(sql);
+				ps.setString(1, login);
+				ps.setLong(2, idUserLogado);
+				rs = ps.executeQuery();
+				while (rs.next()) {
+					ml = new ModelLogin();
+					ml.setId(rs.getLong("id"));
+					ml.setNome(rs.getString("nome"));
+					ml.setEmail(rs.getString("email"));
+					ml.setLogin(rs.getString("login"));
+					ml.setSenha(rs.getString("senha"));
+				}
+			} catch (SQLException e) {
+				Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
+				e.printStackTrace();
+			} finally {
+				if (rs != null)
+					rs.close();
+				if (ps != null)
+					ps.close();
 			}
-		} catch (SQLException e) {
-			e.printStackTrace();
-		} finally {
-			if (rs != null)
-				rs.close();
-			if (ps != null)
-				ps.close();
 		}
 		return ml;
 	}
 	
-	public ModelLogin obterUsuario(Long id) throws Exception {
+	public ModelLogin obterUsuario(String login) throws Exception {
 		ModelLogin ml = null;
-		String sql = "select * from " + TABELA + " where id = ?";
+		String sql = "select * from " + TABELA + " where login = ?"; // login
+																												// é
+																												// único
 		PreparedStatement ps = null;
+		ResultSet rs = null;
+		if (login != null && !login.isEmpty()) {
+			try {
+				ps = connection.prepareStatement(sql);
+				ps.setString(1, login);
+				rs = ps.executeQuery();
+				while (rs.next()) {
+					ml = new ModelLogin();
+					ml.setId(rs.getLong("id"));
+					ml.setNome(rs.getString("nome"));
+					ml.setEmail(rs.getString("email"));
+					ml.setLogin(rs.getString("login"));
+					ml.setSenha(rs.getString("senha"));
+				}
+			} catch (SQLException e) {
+				Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
+				e.printStackTrace();
+			} finally {
+				if (rs != null)
+					rs.close();
+				if (ps != null)
+					ps.close();
+			}
+		}
+		return ml;
+	}
+
+	public ModelLogin obterUsuario(Long id, Long idUserLogado) throws Exception {
+		ModelLogin ml = null;
+		String sql = "select * from " + TABELA + " where id = ? and useradmin is false and usuario_id = ?";
+		PreparedStatement ps = null;
+		ResultSet rs = null;
+
+		if (id != null && idUserLogado != null) {
+			try {
+				ps = connection.prepareStatement(sql);
+				ps.setLong(1, id);
+				ps.setLong(2, idUserLogado);
+				rs = ps.executeQuery();
+				while (rs.next()) {
+					ml = new ModelLogin();
+					ml.setId(rs.getLong("id"));
+					ml.setNome(rs.getString("nome"));
+					ml.setEmail(rs.getString("email"));
+					ml.setLogin(rs.getString("login"));
+					// ml.setSenha(rs.getString("senha"));
+				}
+			} catch (SQLException e) {
+				Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
+				e.printStackTrace();
+			} finally {
+				if (rs != null)
+					rs.close();
+				if (ps != null)
+					ps.close();
+			}
+		}
+		return ml;
+	}
+
+	public List<ModelLogin> consultaUsuarioList(String nome, Long idUserLogado) throws Exception {
+		List<ModelLogin> loginLista = new ArrayList<ModelLogin>();
+		String sql = "select id, nome, email, login, senha from " + TABELA
+				+ " where lower(nome) like concat('%',lower(?),'%') and usuario_id = ? and not useradmin order by nome";
+		PreparedStatement ps = null;
+		ResultSet rs = null;
+		if (nome != null && !nome.isEmpty() && idUserLogado != null) {
+			try {
+				ps = connection.prepareStatement(sql);
+				ps.setString(1, nome);
+				ps.setLong(2, idUserLogado);
+				// ps.setString(1, "%"+ nome +"%"); //não usar o concat
+				rs = ps.executeQuery();
+				while (rs.next()) {
+					ModelLogin ml = new ModelLogin();
+					ml.setId(rs.getLong("id"));
+					ml.setNome(rs.getString("nome"));
+					ml.setEmail(rs.getString("email"));
+					ml.setLogin(rs.getString("login"));
+					// ml.setSenha(rs.getString("senha"));
+					loginLista.add(ml);
+				}
+			} catch (SQLException e) {
+				Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
+				e.printStackTrace();
+			} finally {
+				if (rs != null)
+					rs.close();
+				if (ps != null)
+					ps.close();
+			}
+		}
+
+		return loginLista;
+	}
+
+	public List<ModelLogin> getTodosUsuarios(Long idUserLogado) throws Exception {
+		List<ModelLogin> loginLista = new ArrayList<ModelLogin>();
+		String sql = "select id, nome, email, login, senha from " + TABELA + " where usuario_id = ? and useradmin is false order by id";
+		PreparedStatement ps = null; // Era statment
 		ResultSet rs = null;
 
 		try {
 			ps = connection.prepareStatement(sql);
-			ps.setLong(1, id);
+			ps.setLong(1, idUserLogado);
 			rs = ps.executeQuery();
 			while (rs.next()) {
-				ml = new ModelLogin();
+				ModelLogin ml = new ModelLogin();
 				ml.setId(rs.getLong("id"));
 				ml.setNome(rs.getString("nome"));
 				ml.setEmail(rs.getString("email"));
 				ml.setLogin(rs.getString("login"));
-				//ml.setSenha(rs.getString("senha"));
+				// ml.setSenha(rs.getString("senha"));
+				loginLista.add(ml);
 			}
 		} catch (SQLException e) {
+			Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
 			e.printStackTrace();
 		} finally {
 			if (rs != null)
@@ -97,63 +211,7 @@ public class DaoUsuarioRepository {
 			if (ps != null)
 				ps.close();
 		}
-		return ml;
-	}	
-	
-	public List<ModelLogin> consultaUsuarioList(String nome) throws Exception{
-		List<ModelLogin> loginLista = new ArrayList<ModelLogin>();
-		String sql = "select id, nome, email, login, senha from "+ TABELA +" where lower(nome) like concat('%',lower(?),'%') and not useradmin order by nome";
-		PreparedStatement ps = null;
-		ResultSet rs = null;
-		try {
-			ps = connection.prepareStatement(sql);
-			ps.setString(1, nome);
-			//ps.setString(1, "%"+ nome +"%"); //não usa o concat
-			rs = ps.executeQuery();
-			while (rs.next()) {
-				ModelLogin ml = new ModelLogin();
-				ml.setId(rs.getLong("id"));
-				ml.setNome(rs.getString("nome"));
-				ml.setEmail(rs.getString("email"));
-				ml.setLogin(rs.getString("login"));
-				//ml.setSenha(rs.getString("senha"));
-				loginLista.add(ml);
-			}
-		} catch (SQLException e) {
-			e.printStackTrace();
-		} finally {
-			if (rs != null) rs.close();
-			if (ps != null) ps.close();
-		}
-		
-		return loginLista;
-	}
-	
-	public List<ModelLogin> getTodosUsuarios() throws Exception{
-		List<ModelLogin> loginLista = new ArrayList<ModelLogin>();
-		String sql = "select id, nome, email, login, senha from "+ TABELA +" where not useradmin order by id";
-		Statement stm = null;
-		ResultSet rs = null;
-		try {
-			stm = connection.createStatement();
-			//ps.setString(1, "%"+ nome +"%"); //não usa o concat
-			rs = stm.executeQuery(sql);
-			while (rs.next()) {
-				ModelLogin ml = new ModelLogin();
-				ml.setId(rs.getLong("id"));
-				ml.setNome(rs.getString("nome"));
-				ml.setEmail(rs.getString("email"));
-				ml.setLogin(rs.getString("login"));
-				//ml.setSenha(rs.getString("senha"));
-				loginLista.add(ml);
-			}
-		} catch (SQLException e) {
-			e.printStackTrace();
-		} finally {
-			if (rs != null) rs.close();
-			if (stm != null) stm.close();
-		}
-		
+
 		return loginLista;
 	}
 
@@ -174,7 +232,9 @@ public class DaoUsuarioRepository {
 				connection.commit();
 				atualizado = true;
 			} catch (SQLException e) {
+				Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
 				e.printStackTrace();
+				connection.rollback();
 			} finally {
 				if (ps != null)
 					ps.close();
@@ -182,12 +242,11 @@ public class DaoUsuarioRepository {
 		}
 		return atualizado;
 	}
-	
-	
 
 	public boolean deletarUser(String id) throws Exception {
 		boolean excluiu = false;
-		String sql = "delete from " + TABELA + " where id = ?";
+		String sql = "delete from " + TABELA + " where id = ? and useradmin is false"; // não permite deletar user tipo
+																						// admin
 		PreparedStatement ps = null;
 		Long idUser;
 		if (id != null && !id.isEmpty()) {
@@ -200,6 +259,7 @@ public class DaoUsuarioRepository {
 				connection.commit();
 				excluiu = true;
 			} catch (SQLException e) {
+				Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
 				e.printStackTrace();
 				connection.rollback();
 			} finally {
@@ -224,6 +284,7 @@ public class DaoUsuarioRepository {
 				existe = true;
 			}
 		} catch (SQLException e) {
+			Utilitario.localSqlExceptionMetodo(new Object(){}.getClass().getName(), new Object(){}.getClass().getEnclosingMethod().getName());
 			e.printStackTrace();
 		} finally {
 			if (rs != null)
@@ -234,4 +295,6 @@ public class DaoUsuarioRepository {
 
 		return existe;
 	}
+	
+
 }
